requirements.txt
kafka-python==2.0.2
requests==2.31.0
python-dotenv==1.0.0
alpha-vantage==2.3.1
finnhub-python==2.4.25
praw==7.7.1
newsapi-python==0.2.7
pandas==2.1.3
apscheduler==3.10.4